feat: try to refact render_r1cs
