+++ algebraic/src/lib.rs
@@ -15,7 +15,6 @@ pub mod circom_circuit;
 pub mod errors;
 pub mod field_gl;
 pub mod r1cs;
-pub mod r1cs_reader;
 pub mod r1cs_witness;
 
 pub mod utils;
