--- algebraic/src/r1cs_reader.rs
@@ -1,234 +1,0 @@
-use crate::errors::{EigenError, Result};
-use byteorder::{LittleEndian, ReadBytesExt};
-use itertools::Itertools;
-use std::collections::BTreeMap;
-use std::fs::{File, OpenOptions};
-use std::io::{BufReader, Read, Seek};
-use std::str;
-
-use crate::bellman_ce::{
-    kate_commitment::{Crs, CrsForLagrangeForm, CrsForMonomialForm},
-    pairing::Engine,
-    Field, PrimeField, PrimeFieldRepr, ScalarEngine,
-};
-
-use crate::circom_circuit::CircuitJson;
-use crate::r1cs::r1cs_file::R1CSFile;
-use crate::r1cs::R1CS;
-
-/// get universal setup file by filename
-fn get_universal_setup_file_buff_reader(setup_file_name: &str) -> Result<BufReader<File>> {
-    let setup_file = File::open(setup_file_name).map_err(|e| {
-        EigenError::from(format!(
-            "Failed to open universal setup file {}, err: {}",
-            setup_file_name, e
-        ))
-    })?;
-    Ok(BufReader::with_capacity(1 << 29, setup_file))
-}
-
-/// load monomial form SRS by filename
-pub fn load_key_monomial_form<E: Engine>(filename: &str) -> Crs<E, CrsForMonomialForm> {
-    let mut buf_reader =
-        get_universal_setup_file_buff_reader(filename).expect("read key_monomial_form file err");
-    Crs::<E, CrsForMonomialForm>::read(&mut buf_reader).expect("read key_monomial_form err")
-}
-
-/// load optional lagrange form SRS by filename
-pub fn maybe_load_key_lagrange_form<E: Engine>(
-    option_filename: Option<String>,
-) -> Option<Crs<E, CrsForLagrangeForm>> {
-    match option_filename {
-        None => None,
-        Some(filename) => {
-            let mut buf_reader = get_universal_setup_file_buff_reader(&filename)
-                .expect("read key_lagrange_form file err");
-            let key_lagrange_form = Crs::<E, CrsForLagrangeForm>::read(&mut buf_reader)
-                .expect("read key_lagrange_form err");
-            Some(key_lagrange_form)
-        }
-    }
-}
-
-/// load r1cs_witness file by filename with autodetect encoding (bin or json).
-pub fn load_witness_from_file<E: ScalarEngine>(filename: &str) -> Vec<E::Fr> {
-    if filename.ends_with("json") {
-        load_witness_from_json_file::<E>(filename)
-    } else {
-        load_witness_from_bin_file::<E>(filename)
-    }
-}
-
-/// load r1cs_witness from json file by filename
-pub fn load_witness_from_json_file<E: ScalarEngine>(filename: &str) -> Vec<E::Fr> {
-    let reader = OpenOptions::new()
-        .read(true)
-        .open(filename)
-        .expect("unable to open.");
-    load_witness_from_json::<E, BufReader<File>>(BufReader::new(reader))
-}
-
-/// load r1cs_witness from json by a reader
-fn load_witness_from_json<E: ScalarEngine, R: Read>(reader: R) -> Vec<E::Fr> {
-    let witness: Vec<String> = serde_json::from_reader(reader).expect("unable to read.");
-    witness
-        .into_iter()
-        .map(|x| E::Fr::from_str(&x).unwrap())
-        .collect::<Vec<E::Fr>>()
-}
-
-/// load r1cs_witness from bin file by filename
-pub fn load_witness_from_bin_file<E: ScalarEngine>(filename: &str) -> Vec<E::Fr> {
-    let reader = OpenOptions::new()
-        .read(true)
-        .open(filename)
-        .expect("unable to open.");
-    load_witness_from_bin_reader::<E, BufReader<File>>(BufReader::new(reader))
-        .expect("read r1cs_witness failed")
-}
-
-/// load r1cs_witness from u8 array
-pub fn load_witness_from_array<E: ScalarEngine>(buffer: Vec<u8>) -> Result<Vec<E::Fr>> {
-    load_witness_from_bin_reader::<E, _>(buffer.as_slice())
-}
-
-/// load r1cs_witness from u8 array by a reader
-pub fn load_witness_from_bin_reader<E: ScalarEngine, R: Read>(mut reader: R) -> Result<Vec<E::Fr>> {
-    let mut wtns_header = [0u8; 4];
-    reader.read_exact(&mut wtns_header)?;
-    if wtns_header != [119, 116, 110, 115] {
-        // python -c 'print([ord(c) for c in "wtns"])' => [119, 116, 110, 115]
-        return Err(EigenError::from("Invalid file header".to_string()));
-    }
-    let version = reader.read_u32::<LittleEndian>()?;
-    log::debug!("wtns version {}", version);
-    if version > 2 {
-        return Err(EigenError::from("unsupported file version".to_string()));
-    }
-    let num_sections = reader.read_u32::<LittleEndian>()?;
-    if num_sections != 2 {
-        return Err(EigenError::from("invalid num sections".to_string()));
-    }
-    // read the first section
-    let sec_type = reader.read_u32::<LittleEndian>()?;
-    if sec_type != 1 {
-        return Err(EigenError::from("invalid section type".to_string()));
-    }
-    let sec_size = reader.read_u64::<LittleEndian>()?;
-    if sec_size != 4 + 32 + 4 {
-        return Err(EigenError::from("invalid section len".to_string()));
-    }
-    let field_size = reader.read_u32::<LittleEndian>()?;
-    if field_size != 32 {
-        return Err(EigenError::from("invalid field byte size".to_string()));
-    }
-    let mut prime = vec![0u8; field_size as usize];
-    reader.read_exact(&mut prime)?;
-    if prime != hex!("010000f093f5e1439170b97948e833285d588181b64550b829a031e1724e6430") {
-        return Err(EigenError::from("invalid curve prime".to_string()));
-    }
-    let witness_len = reader.read_u32::<LittleEndian>()?;
-    log::debug!("r1cs_witness len {}", witness_len);
-    let sec_type = reader.read_u32::<LittleEndian>()?;
-    if sec_type != 2 {
-        return Err(EigenError::from("invalid section type".to_string()));
-    }
-    let sec_size = reader.read_u64::<LittleEndian>()?;
-    if sec_size != (witness_len * field_size) as u64 {
-        return Err(EigenError::from(format!(
-            "Invalid r1cs_witness section size {}",
-            sec_size
-        )));
-    }
-    let mut result = Vec::with_capacity(witness_len as usize);
-    for _ in 0..witness_len {
-        let mut repr = E::Fr::zero().into_repr();
-        repr.read_le(&mut reader)?;
-        result.push(E::Fr::from_repr(repr)?);
-    }
-    Ok(result)
-}
-
-/// load r1cs file by filename with autodetect encoding (bin or json)
-pub fn load_r1cs<E: ScalarEngine>(filename: &str) -> R1CS<E> {
-    if filename.ends_with("json") {
-        load_r1cs_from_json_file(filename)
-    } else {
-        let (r1cs, _wire_mapping) = load_r1cs_from_bin_file(filename);
-        r1cs
-    }
-}
-
-/// load r1cs from json file by filename
-fn load_r1cs_from_json_file<E: ScalarEngine>(filename: &str) -> R1CS<E> {
-    let reader = OpenOptions::new()
-        .read(true)
-        .open(filename)
-        .expect("unable to open.");
-    load_r1cs_from_json(BufReader::new(reader))
-}
-
-/// load r1cs from json by a reader
-fn load_r1cs_from_json<E: ScalarEngine, R: Read>(reader: R) -> R1CS<E> {
-    let circuit_json: CircuitJson = serde_json::from_reader(reader).expect("unable to read.");
-
-    let num_inputs = circuit_json.num_inputs + circuit_json.num_outputs + 1;
-    let num_aux = circuit_json.num_variables - num_inputs;
-
-    let convert_constraint = |lc: &BTreeMap<String, String>| {
-        lc.iter()
-            .map(|(index, coeff)| (index.parse().unwrap(), E::Fr::from_str(coeff).unwrap()))
-            .collect_vec()
-    };
-
-    let constraints = circuit_json
-        .constraints
-        .iter()
-        .map(|c| {
-            (
-                convert_constraint(&c[0]),
-                convert_constraint(&c[1]),
-                convert_constraint(&c[2]),
-            )
-        })
-        .collect_vec();
-
-    R1CS {
-        num_inputs,
-        num_aux,
-        num_variables: circuit_json.num_variables,
-        num_outputs: circuit_json.num_outputs,
-        constraints,
-        custom_gates: vec![],
-        custom_gates_uses: vec![],
-    }
-}
-
-/// load r1cs from bin file by filename
-fn load_r1cs_from_bin_file<E: ScalarEngine>(filename: &str) -> (R1CS<E>, Vec<usize>) {
-    let reader = OpenOptions::new()
-        .read(true)
-        .open(filename)
-        .unwrap_or_else(|_| panic!("unable to open {}.", filename));
-    load_r1cs_from_bin(BufReader::new(reader))
-}
-
-/// load r1cs from bin by a reader
-pub fn load_r1cs_from_bin<R: Read + Seek, E: ScalarEngine>(reader: R) -> (R1CS<E>, Vec<usize>) {
-    let file = R1CSFile::<E>::from_reader::<R>(reader).expect("unable to read.");
-    let num_inputs = (1 + file.header.n_pub_in + file.header.n_pub_out) as usize;
-    let num_variables = file.header.n_wires as usize;
-    let num_aux = num_variables - num_inputs;
-    (
-        R1CS {
-            num_aux,
-            num_inputs,
-            num_variables,
-            num_outputs: file.header.n_pub_out as usize,
-            constraints: file.constraints,
-            custom_gates: file.custom_gates,
-            custom_gates_uses: file.custom_gates_uses,
-        },
-        file.wire_mapping.iter().map(|e| *e as usize).collect_vec(),
-    )
-}
