+++ starky/src/r1cs2plonk.rs
@@ -252,7 +252,6 @@ pub fn r1cs2plonk(r1cs: &R1CS<GL>) -> (Vec<PlonkGate>, Vec<PlonkAdd>) {
 #[cfg(test)]
 mod test {
     use super::*;
-    use algebraic::r1cs_reader::load_r1cs;
     use std::fs::File;
     use std::io::Write;
     use std::path::Path;
